requirements/base.txt

py3-validate-email==1.0.2
python-dateutil>=2.7.3
confuse>=1.3.0
pyyaml>=5.3.1
errorhandler>=2.0.1
lxml>=4.6.1